Tunnels: Move 3 build classes from tunnel to tunnel/pool
and make package private, so all the tunnel build classes are in one place.

(cherry picked from commit 4f12081b59858207fc622868225015b240f71228)
